Cygwin: fhandler_fifo::close: fix a race condition
When the owning reader closes, there's a loop in which it tries to
find a new owner.  It decrements the reader count at the beginning of
the loop and increments it at the end.  If another reader closes
during this time, the inaccurate reader count might cause the
read_ready event to get reset, thereby blocking a writer that wants to
open.

Fix this by introducing a new method, fhandler_fifo::nreaders that
returns the number of readers.  Use this instead of decrementing and
incrementing the reader count.